refactor(Services): consumir novos métodos dos repositorios.

+++ rede-social/rede-social-service/src/main/java/br/com/crescer/service/PessoaService.java
@@ -2,7 +2,7 @@ package br.com.crescer.service;
 
 import br.com.crescer.entity.Pessoa;
 import br.com.crescer.repository.PessoaRepository;
-import java.math.BigDecimal;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -15,7 +15,7 @@ public class PessoaService {
     @Autowired
     PessoaRepository repository;
 
-    public Pessoa getPessoa(BigDecimal id) {
+    public Pessoa getPessoa(Long id) {
         return repository.findOne(id);
     }
 }
